Move unit tests to their own unit directory
This will make it easier to run unit tests separately from
integration tests, which will require a running DB.
