+++ test/unit/utils/form.test.ts
@@ -1,5 +1,5 @@
 import { ZodError } from "zod";
-import * as form from "../../app/utils/form";
+import * as form from "../../../app/utils/form";
 
 describe("formatValidationErrors", () => {
   const error = new ZodError([
